requirements.txt
opencv-python
pyzmq
numpy
dash
